JS syntax optimisation
It's been a while since I've maintained this addon, so my JS skills have marginally improved. This means it's syntax and optimisation time!! :3

I'm not sure what this commit will do performance-wise, but ideally it should bring minor stability and \[probably more notably\] byte-count-type improvements. And even then, it's going to be minor if not negligible.