Restructured code inheritance
Manifest module inheritance does work but is super vague. Generally, the order of the scripts specified is the order they're run in and thus that defines who can inherit (only newer scripts). This is kinda annoying in some cases where inheritance needed isn't exactly linear (for example, in configurator.js needing `log()`); so, I've created common.js, which is run first.

I'll add functions and globals here when needed.

Additionally, popup.html has to specify this one manually first given it's out of scope of the manifest. But that's okay.