New Travel/Flights selectors; new Travel favicon
Fixes #25.

It's a really horrible hack to just append garbage onto the `href` of the favicon elements, but I think Google is using some nasty nasty deferment. If I set the `href` property to `` or remove the `<link>` element for the favicons it provides automatically, it doesn't work. _Only_ mangling the `href` works.