Fixed possible failure in `getConfig()`
Looking up an invalid key in an object doesn't return a TypeError (like it would've used to in the old array data structure, where an invalid key returned an array index of -1: out of bounds). So instead, if the yielded value is null (unset), we know it doesn't exist, so return `false`.