Disabled debugging; getCaller() patch
Debugging was on in production :(

Proper fix for v3.1.3 bugfix

Original patch attempt was in 67e810e2300d1f96468c7ca8c977fff3b9c11c6a. getCaller() still gets called in debug and is subject to the same issue, causing the same crash. I _believe_ it was caused by an out of bounds when you call \`getCaller(3)\` or something, but the stack isn't that long so you get errors like \"error[i] is undefined\". That's my theory.

Fixed by manually checking that the provided stack depth cannot be larger than the final index in \`error\`.